http://issues.apache.org/activemq/browse/AMQCPP-103
Adding in more Types wrappers

git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@560036 13f79535-47bb-0310-9956-ffa450edef68

+++ src/decaf/src/main/decaf/lang/Number.h
@@ -63,9 +63,9 @@ namespace lang{
 
         /**
          * Answers the long value which the receiver represents
-         * @return long the value of the receiver.
+         * @return long long the value of the receiver.
          */
-        virtual long longValue() const = 0;
+        virtual long long longValue() const = 0;
 
         /**
          * Answers the short value which the receiver represents
